Renaming to match diagrams
