+++ Project/src/app/app.component.ts
@@ -1,8 +1,8 @@
 import { Component, ViewChild } from '@angular/core';
 import { Nav, Platform } from 'ionic-angular';
 import { StatusBar, Splashscreen } from 'ionic-native';
-import { PhonemeList } from '../pages/PhonemeList/PhonemeList';
-import { ProfileSetup } from '../pages/ProfileSetup/ProfileSetup';
+import { LessonsList } from '../pages/LessonsList/LessonsList';
+import { ProfileManager } from '../pages/ProfileManager/ProfileManager';
 import { ProfileInfo } from '../loaders/profileInfo';
 
 @Component({
@@ -28,18 +28,18 @@ export class Phonetics {
             user: this.user
           };
 
-          this.nav.setRoot(PhonemeList, params);
+          this.nav.setRoot(LessonsList, params);
         } else {
-          this.rootPage = ProfileSetup;
+          this.rootPage = ProfileManager;
         }
       })
-      .catch(err => this.rootPage = ProfileSetup);
+      .catch(err => this.rootPage = ProfileManager);
 
 
     // used for an example of ngFor and navigation
     this.pages = [
-      { title: 'Lessons', component: PhonemeList },
-      { title: 'Edit Profile', component: ProfileSetup }
+      { title: 'Lessons', component: LessonsList },
+      { title: 'Edit Profile', component: ProfileManager }
     ];
   }
 
@@ -57,7 +57,7 @@ export class Phonetics {
     // Reset the content nav to have just this page
     // we wouldn't want the back button to show in this scenario
     let params;
-    if (page.component === ProfileSetup || page.component === PhonemeList) {
+    if (page.component === ProfileManager || page.component === LessonsList) {
       if (this.user) {
         params = {
           user: this.user
