+++ Project/src/pages/LessonsList/LessonsList.scss
@@ -1,4 +1,4 @@
-page-phonemeList {
+page-lessonsList {
     a {
         color: map-get($colors, dark);
     }
