+++ Project/src/pages/ListeningController/ListeningController.scss
@@ -1,4 +1,4 @@
-page-SpeakingMode {
+page-ListeningController {
     ion-card {
         position: absolute;
         top: 0;
