+++ Project/src/pages/ProfileManager/ProfileManager.scss
@@ -1,4 +1,4 @@
-page-ProfileSetup {
+page-ProfileManager {
     ion-card {
         position: absolute;
         top: 0;
