+++ Project/src/pages/SpeakingController/SpeakingController.scss
@@ -1,4 +1,4 @@
-page-ListeningMode {
+page-SpeakingController {
     ion-card {
         position: absolute;
         top: 0;
